Separate Azure membership into a separate library and move shared pieces to Azure.Core
Move shared classes to a central shared folder instead of separate project

Address feedback
