+++ Orleans.sln
@@ -138,11 +138,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Transactions.Azure"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Transactions.Azure.Test", "test\Orleans.Transactions.Azure.Test\Orleans.Transactions.Azure.Test.csproj", "{651C7B8E-6EEB-4A6B-84A3-B5D7E4554B99}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Clustering.AzureStorage", "src\Orleans.Clustering.AzureStorage\Orleans.Clustering.AzureStorage.csproj", "{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}"
+EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.Hosting.ServiceFabric", "src\Orleans.Hosting.ServiceFabric\Orleans.Hosting.ServiceFabric.csproj", "{1AF48B10-B2CE-4DE9-9A3F-0A656FBF1C02}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "ServiceFabric", "ServiceFabric", "{A15BD3B2-513B-449C-A2E9-81DA2DE2FF92}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Orleans.ServiceFabric", "src\Orleans.ServiceFabric\Orleans.ServiceFabric.csproj", "{814FF01E-1466-4A71-A48F-F40143EF49AA}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Orleans.ServiceFabric", "src\Orleans.ServiceFabric\Orleans.ServiceFabric.csproj", "{814FF01E-1466-4A71-A48F-F40143EF49AA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -898,6 +900,18 @@ Global
 		{651C7B8E-6EEB-4A6B-84A3-B5D7E4554B99}.Release|x64.Build.0 = Release|Any CPU
 		{651C7B8E-6EEB-4A6B-84A3-B5D7E4554B99}.Release|x86.ActiveCfg = Release|Any CPU
 		{651C7B8E-6EEB-4A6B-84A3-B5D7E4554B99}.Release|x86.Build.0 = Release|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Debug|x64.Build.0 = Debug|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Debug|x86.Build.0 = Debug|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Release|x64.ActiveCfg = Release|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Release|x64.Build.0 = Release|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Release|x86.ActiveCfg = Release|Any CPU
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76}.Release|x86.Build.0 = Release|Any CPU
 		{1AF48B10-B2CE-4DE9-9A3F-0A656FBF1C02}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{1AF48B10-B2CE-4DE9-9A3F-0A656FBF1C02}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1AF48B10-B2CE-4DE9-9A3F-0A656FBF1C02}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -992,6 +1006,7 @@ Global
 		{FE2E08C6-9C3B-4AEE-AE07-CCA387580D7A} = {4CD3AA9E-D937-48CA-BB6C-158E12257D23}
 		{BE7FB595-FA25-4D88-8504-E9D4F9D8183D} = {FE2E08C6-9C3B-4AEE-AE07-CCA387580D7A}
 		{651C7B8E-6EEB-4A6B-84A3-B5D7E4554B99} = {A6573187-FD0D-4DF7-91D1-03E07E470C0A}
+		{E9A8A1F9-1943-45E2-9B9C-1E51A9154D76} = {FE2E08C6-9C3B-4AEE-AE07-CCA387580D7A}
 		{1AF48B10-B2CE-4DE9-9A3F-0A656FBF1C02} = {A15BD3B2-513B-449C-A2E9-81DA2DE2FF92}
 		{A15BD3B2-513B-449C-A2E9-81DA2DE2FF92} = {FE2E08C6-9C3B-4AEE-AE07-CCA387580D7A}
 		{814FF01E-1466-4A71-A48F-F40143EF49AA} = {A15BD3B2-513B-449C-A2E9-81DA2DE2FF92}
