+++ src/Orleans.Clustering.AzureStorage/LegacyAzureMembershipConfigurator.cs
