+++ src/Orleans.Clustering.AzureStorage/Options/AzureTableGatewayListProviderOptions.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace OrleansAzureUtils.Options
+﻿namespace Orleans.AzureUtils.Options
 {
     public class AzureTableGatewayListProviderOptions
     {
