+++ src/Orleans.Clustering.AzureStorage/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+using System.Runtime.CompilerServices;
+
+// TODO Temporary until we totally remove Orleans.AzureUtils
+[assembly: InternalsVisibleTo("OrleansAzureUtils")]
+
+[assembly: InternalsVisibleTo("Tester.AzureUtils")]
