+++ src/Orleans.Core/Messaging/ILegacyGatewayListProviderConfigurator.cs
@@ -30,7 +30,7 @@ namespace Orleans.Messaging
             {
                 case ClientConfiguration.GatewayProviderType.AzureTable:
                     configurator = CreateInstanceWithParameterlessConstructor<ILegacyGatewayListProviderConfigurator>(
-                            Constants.ORLEANS_AZURE_UTILS_DLL);
+                            Constants.ORLEANS_CLUSTERING_AZURESTORAGE);
                     break;
 
                 case ClientConfiguration.GatewayProviderType.SqlServer:
