+++ test/TesterAzureUtils/AzureMembershipTableTests.cs
@@ -5,11 +5,11 @@ using Microsoft.Extensions.Options;
 using Orleans;
 using Orleans.AzureUtils;
 using Orleans.AzureUtils.Configuration;
+using Orleans.AzureUtils.Options;
 using Orleans.Messaging;
 using Orleans.Runtime;
 using Orleans.Runtime.Configuration;
 using Orleans.Runtime.MembershipService;
-using OrleansAzureUtils.Options;
 using TestExtensions;
 using UnitTests;
 using UnitTests.MembershipTests;
